+++ packages/coinstac-images/coinstac-file-transfer-test/test/input/local0/simulatorRun/b.txt
@@ -1,0 +1,1 @@
+bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
